fixed player taking dmg while dash

+++ Assets/Objects/Scripts/Bullet_Settings.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Bullet_Settings : ScriptableObject
 {
     public float speed = 10f;
-    public float baseDamage = 2f;
+    public int baseDamage = 2;
     public Vector3 scale = new Vector3(1f, 1f, 1f);
     [HideInInspector] public Vector3 target;
 }
